use clip for bounding box transformation and correct the dtype while mul op

+++ keras/src/layers/preprocessing/image_preprocessing/bounding_boxes/validation.py
@@ -110,7 +110,7 @@ def densify_bounding_boxes(
         )
         return bounding_boxes
 
-    bounding_boxes["boxes"] = backend.convert_to_tensor(boxes, dtype="int32")
+    bounding_boxes["boxes"] = backend.convert_to_tensor(boxes, dtype="float32")
     bounding_boxes["labels"] = backend.convert_to_tensor(labels)
     return bounding_boxes
